refactor useTitleFormatter() hook
