+++ packages/docusaurus-theme-common/src/internal.ts
@@ -43,7 +43,7 @@ export {
 
 export {DEFAULT_SEARCH_TAG} from './utils/searchUtils';
 
-export {useTitleFormatter} from './utils/generalUtils';
+export {useTitleFormatter} from './utils/titleFormatterUtils';
 
 export {useLocationChange} from './utils/useLocationChange';
